requirements.txt
# Project pinned requirements (tested with Python 3.12)
duckdb==1.1.1
pandas==2.3.3
pyarrow==21.0.0
numpy==2.3.3
altair==5.5.0
streamlit==1.50.0
python-dotenv==1.1.1
requests==2.32.5
holidays==0.81
sodapy==2.2.0
pytest==8.4.2
